+++ i18n/_messages.en.js
@@ -17,6 +17,7 @@ window.ParsleyConfig = window.ParsleyConfig || {};
           , digits:     "This value should be digits."
           , dateIso:    "This value should be a valid date (YYYY-MM-DD)."
           , alphanum:   "This value should be alphanumeric."
+          , phone:      "This value should be a valid phone number."
         }
       , notnull:        "This value should not be null."
       , notblank:       "This value should not be blank."
@@ -34,8 +35,7 @@ window.ParsleyConfig = window.ParsleyConfig || {};
       , equalto:        "This value should be the same."
 
       // parsley.extend ///////////////////////////////
-      , minwords:       "Aquest valor ha de tenir %s paraules com a mínim."
-        minwords:       "This value should have %s words at least."
+      , minwords:       "This value should have %s words at least."
       , maxwords:       "This value should have %s words maximum."
       , rangewords:     "This value should have between %s and %s words."
       , greaterthan:    "This value should be greater than %s."